Make the socket non-blocking and connect asychronously
I was unable to use this module. When calling the `connect` method,
asyncio was complaining that the socket must be non-blocking (might have
to enable asyncio debugging to get the error, otherwise it will silently
do nothing).

Use asyncio's sock_connect method instead of manually connecting the
socket and make the socket non-blocking as required. To make this fully
work, I had to make the `connect` method async all the way to the top.

With these changes, I can use this module:
- in an interactive IPython shell or a jupyter notebook like so:

    import openspace
    import asyncio

    ADDRESS = 'localhost'
    PORT = 4681

    api = openspace.Api(ADDRESS, PORT)
    await api.connect()
    os = await api.singleReturnLibrary()
    await os.time.togglePause()

- in a regular script like so:

        #! /usr/bin/env python3

        import openspace
        import asyncio

        ADDRESS = 'localhost'
        PORT = 4681

        api = openspace.Api(ADDRESS, PORT)

        async def init():
            await api.connect()
            os = await api.singleReturnLibrary()
            return os

        async def main():
            os = await init()
            await os.time.togglePause()

        asyncio.run(main())